Feat: mobile header menu
