+++ src/components/Global/MenuData.js
@@ -1,15 +1,20 @@
+import React from 'react';
+import * as AiIcons from 'react-icons/ai';
+
 export const MenuData = [
   {
     id: 1,
     type: 'text',
     title: '선거정보',
     path: '/info',
+    icon: <AiIcons.AiOutlineFile />,
   },
   {
     id: 2,
     type: 'text',
     title: '공약게시판',
     path: '/board',
+    icon: <AiIcons.AiOutlineOrderedList />,
   },
   {
     id: 3,
@@ -20,7 +25,8 @@ export const MenuData = [
   {
     id: 4,
     type: 'text',
-    title: '관리자게시판',
+    title: '관리자페이지',
     path: '/admin',
+    icon: <AiIcons.AiOutlineBank />,
   },
 ];
