+++ src/lib/styles/theme.js
@@ -1,6 +1,7 @@
 const theme = {
   DarkBlue: '#252c44',
   Blue: '#5d3fe8',
+  Shadow: '#968dbf',
 };
 
 export default theme;
